refactor: 현재 사용자 정보 조회 함수로 분리
